+++ src/routes.js
@@ -7,7 +7,7 @@ import DeepSearchPage from './components/DeepSearchPage/';
 export default () => (
   <HashRouter>
     <Switch>
-      <Route path='/search/deep' component={DeepSearchPage} />
+      <Route path='/deepsearch/:search' component={DeepSearchPage} />
       <Route path='/:search' component={FoamTreeSearchPage} />
       <Route path='/' component={LandingPage} />
     </Switch>
